Update README.md
In this video I will show you how to build your own web scraper for Yellow Pages or Yell. We can easily extract a lot of useful leads and business data by using a simple web scraper. We use requests and BeautifulSoup to download and parse the data and export it to CSV for Excel